requirements.txt
ruamel.yaml<0.15
simplejson
appdirs